Our Replication of The Cash Paradox (Jiang & Shao, 2019)

This replication study was part of our evaluation for the course Numerical Methods at SciencesPo Paris in Spring 2021

In this replication study, we do ...

CashParadox.load_data — Type

load_data(irflag,flag)

Load data depending on the figure and country wanted.

CashParadox.eqn_LWfit — Method

eqn_LWfit

Matching the time series of CIC_GDP ratio in the data with the simulated series in the full model

CashParadox.eqn_Regime201610 — Method

eqn_Regime201610

Extensive margin Equations to solve q2 and y in regime 2

CashParadox.eqn_noconnfit — Method

eqn_noconnfit

Matching the time series of CIC_GDP ratio in the data with the simulated series in the full model

CashParadox.eqn_tfit — Method

eqn_tfit

Matching the time series of CIC_GDP ratio in the data with the simulated series in the full model

CashParadox.figA3 — Method

figA3(x)

Creates figure A3

CashParadox.figD1 — Method

figD1(x)

Creates figure D1

CashParadox.figD2 — Method

figD4(x)

Creates figure D2

CashParadox.figsub5 — Method

figsub5(irflag,flag,model)

Creates the subplots for figure 5 for a specific interest rate specification and country

CashParadox.figsubA2 — Method

figsubA2(irflag,flag,model)

Creates the subplots for figure 5 for a specific interest rate specification and country

CashParadox.figsubD2 — Method

figsubD2(irflag,flag,model)

Creates subplots of figure D2

CashParadox.fmincon — Function

xopt, fopt, converged = fmincon(obj, startval)

Minimize the function obj, starting at startval. fminunc() with no arguments will run an example, execute edit(fminunc,()) to see the code. fminunc() uses NLopt.jl to do the actual minimization.

CashParadox.vs — Method

vs(irflag,flag,model)

Returns vθ and vρ for a specified model

CashParadox.vδ — Method

vδ(irflag,flag)

Computes vδ

end