Our Replication of The Cash Paradox (Jiang & Shao, 2019)

This replication study was part of our evaluation for the course Numerical Methods at SciencesPo Paris in Spring 2021

In this replication study, we do ...

CashParadox.load_data — Type

load_data(irflag,flag)

Load data depending on the figure and country wanted.

CashParadox.eqn_LWfit — Method

eqn_LWfit(x)

Matching the time series of CIC_GDP ratio in the data with the simulated series in the Lagos-Wright model.

CashParadox.fig5 — Method

fig5(irflag,flag)

Replicates figure 5 for a specific country

end