My Replication of xyz

This replication study was part of my evaluation for the course Numerical Methods at SciencesPo Paris in Spring 2021

In this replication study, we do ...

CashParadox.eqn_LWfit — Method

eqn_LWfit(x)

Matching the time series of CIC_GDP ratio in the data with the simulated series in the Lagos-Wright model.

CashParadox.load_data — Method

load_data(irflag,flag)

Load data depending on the figure and country wanted.

end